backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless>=4.8
numpy<2
pytesseract==0.3.10
python-stdnum==1.19
pymupdf>=1.24
